imessage-bridge/pyproject.toml
[project]
name = "imessage-bridge"
version = "0.1.0"
description = "Bridge iMessage to Nest V2 Agent via Supabase"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0",
    "httpx>=0.27",
    "supabase>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
imessage-bridge = "imessage_bridge.__main__:main"
